\name{PhotoMapper-package}
\alias{PhotoMapper-package}
\alias{PhotoMapper}
\docType{package}
\title{
\packageTitle{PhotoMapper}
}
\description{
\packageDescription{PhotoMapper}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PhotoMapper}
\packageIndices{PhotoMapper}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PhotoMapper}

Maintainer: \packageMaintainer{PhotoMapper}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
